requirements.txt
certifi==2019.6.16
numpy==1.16.4
pandas==0.24.2
Pillow==6.1.0
PyMySQL==0.9.3
pyodbc==4.0.26
pytesseract==0.2.7
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.5
wincertstore==0.2
